+++ _octal.c
@@ -18,7 +18,7 @@ int _octal(va_list args)
 	}
 	while (n > 0)
 	{
-		int tmp = 0;
+		long int tmp = 0;
 
 		tmp = n % 8;
 		if (tmp < 10)
